Update README and Rename the phymanager
