+++ README.md
@@ -1,5 +1,9 @@
 # PDSL
 This is a python-like domain specific language for physical computing. It support fundamental unit computing and checking and a simple vector computing.
+##Install 
+run `pip3 -m requirements.txt` to install the python requirements
 
 ##For Development
+run `pip3 -m requirements-dev.txt` to install the python requirements
+
 Please run `./python3-ci.sh` and prove no error in the ci test before merging in to the master.
