+++ requirements.txt
@@ -1,0 +1,2 @@
+numpy==1.16.2
+typing==3.6.6
